S11 after review. Refactor tests.
